feat: hide groupRowEditable and groupRowValueSetter from public API
- Move groupRowEditable and groupRowValueSetter types to internal colDefInternal.ts
- Remove exports from main.ts
- Remove from public ColDef interface
- Update documentation to remove references to the unfinished feature
- Keep internal functionality working with ColDefInternal type
- Remove group-editable-totals-custom and pivot-editable-totals-custom examples
- Keep refresh-after-group-edit example (this feature stays public)
