--- documentation/ag-grid-docs/src/content/docs/grouping-edit/_examples/group-editable-totals-custom/data.ts
@@ -1,43 +1,0 @@
-export const getData = () => [
-    // Europe - Corporate
-    { id: 'fr-paris', region: 'Europe', segment: 'Corporate', country: 'France', amount: 30 },
-    { id: 'fr-lyon', region: 'Europe', segment: 'Corporate', country: 'France', amount: 30 },
-    { id: 'de-berlin', region: 'Europe', segment: 'Corporate', country: 'Germany', amount: 35 },
-    { id: 'de-hamburg', region: 'Europe', segment: 'Corporate', country: 'Germany', amount: 25 },
-    { id: 'es-madrid', region: 'Europe', segment: 'Corporate', country: 'Spain', amount: 28 },
-    { id: 'es-barcelona', region: 'Europe', segment: 'Corporate', country: 'Spain', amount: 32 },
-    { id: 'uk-london', region: 'Europe', segment: 'Corporate', country: 'UK', amount: 45 },
-    { id: 'uk-manchester', region: 'Europe', segment: 'Corporate', country: 'UK', amount: 35 },
-    // Europe - Enterprise
-    { id: 'it-rome', region: 'Europe', segment: 'Enterprise', country: 'Italy', amount: 40 },
-    { id: 'it-milan', region: 'Europe', segment: 'Enterprise', country: 'Italy', amount: 20 },
-    { id: 'pl-warsaw', region: 'Europe', segment: 'Enterprise', country: 'Poland', amount: 26 },
-    { id: 'pl-krakow', region: 'Europe', segment: 'Enterprise', country: 'Poland', amount: 24 },
-    { id: 'nl-amsterdam', region: 'Europe', segment: 'Enterprise', country: 'Netherlands', amount: 38 },
-    { id: 'nl-rotterdam', region: 'Europe', segment: 'Enterprise', country: 'Netherlands', amount: 32 },
-    // Americas - Corporate
-    { id: 'us-nyc', region: 'Americas', segment: 'Corporate', country: 'USA', amount: 70 },
-    { id: 'us-la', region: 'Americas', segment: 'Corporate', country: 'USA', amount: 30 },
-    { id: 'us-austin', region: 'Americas', segment: 'Corporate', country: 'USA', amount: 25 },
-    { id: 'us-chicago', region: 'Americas', segment: 'Corporate', country: 'USA', amount: 55 },
-    { id: 'ca-toronto', region: 'Americas', segment: 'Corporate', country: 'Canada', amount: 35 },
-    { id: 'ca-vancouver', region: 'Americas', segment: 'Corporate', country: 'Canada', amount: 25 },
-    // Americas - Enterprise
-    { id: 'br-sao-paulo', region: 'Americas', segment: 'Enterprise', country: 'Brazil', amount: 30 },
-    { id: 'br-rio', region: 'Americas', segment: 'Enterprise', country: 'Brazil', amount: 22 },
-    { id: 'mx-tijuana', region: 'Americas', segment: 'Enterprise', country: 'Mexico', amount: 28 },
-    { id: 'mx-guadalajara', region: 'Americas', segment: 'Enterprise', country: 'Mexico', amount: 18 },
-    { id: 'ar-buenos-aires', region: 'Americas', segment: 'Enterprise', country: 'Argentina', amount: 24 },
-    // Asia Pacific - Corporate
-    { id: 'jp-tokyo', region: 'Asia Pacific', segment: 'Corporate', country: 'Japan', amount: 65 },
-    { id: 'jp-osaka', region: 'Asia Pacific', segment: 'Corporate', country: 'Japan', amount: 45 },
-    { id: 'au-sydney', region: 'Asia Pacific', segment: 'Corporate', country: 'Australia', amount: 42 },
-    { id: 'au-melbourne', region: 'Asia Pacific', segment: 'Corporate', country: 'Australia', amount: 38 },
-    // Asia Pacific - Enterprise
-    { id: 'cn-shanghai', region: 'Asia Pacific', segment: 'Enterprise', country: 'China', amount: 80 },
-    { id: 'cn-beijing', region: 'Asia Pacific', segment: 'Enterprise', country: 'China', amount: 70 },
-    { id: 'sg-singapore', region: 'Asia Pacific', segment: 'Enterprise', country: 'Singapore', amount: 55 },
-    { id: 'kr-seoul', region: 'Asia Pacific', segment: 'Enterprise', country: 'South Korea', amount: 48 },
-    { id: 'in-mumbai', region: 'Asia Pacific', segment: 'Enterprise', country: 'India', amount: 36 },
-    { id: 'in-bangalore', region: 'Asia Pacific', segment: 'Enterprise', country: 'India', amount: 44 },
-];
