--- documentation/ag-grid-docs/src/content/docs/grouping-edit/_examples/group-editable-totals-custom/example.spec.ts
@@ -1,8 +1,0 @@
-import { ensureGridReady, test, waitForGridContent } from '@utils/grid/test-utils';
-
-test.agExample(import.meta, () => {
-    test.eachFramework('Example', async ({ page }) => {
-        await ensureGridReady(page);
-        await waitForGridContent(page);
-    });
-});
