--- documentation/ag-grid-docs/src/content/docs/grouping-edit/_examples/pivot-editable-totals-custom/index.html
@@ -1,1 +1,0 @@
-<div id="myGrid" class="ag-theme-quartz" style="height: 100%; width: 100%"></div>
