+++ documentation/ag-grid-docs/src/content/docs/pivoting/index.mdoc
@@ -97,10 +97,6 @@ Pivoting can be controlled using the following grid API methods:
 
 {% apiDocumentation source="grid-api/api.json" section="rowPivoting" /%}
 
-## See Also
-
--   [Editing Groups](./grouping-edit/) for editing pivot aggregated values with cascading updates to children
-
 ## Next Up
 
 Continue to the next section to learn about how to customise [Pivot Result Columns](./pivoting-result-columns/).
