+++ documentation/ag-grid-docs/src/content/docs/tree-data-nesting/index.mdoc
@@ -88,8 +88,7 @@ const gridOptions = {
 };
 ```
 
-Refer to the [Aggregation](./aggregation/) page for more details, and [Editing Groups](./grouping-edit/)
-for editing aggregated values with cascading updates to children.
+Refer to the [Aggregation](./aggregation/) page for more details.
 
 ## Next Up
