+++ documentation/ag-grid-docs/src/content/docs/tree-data-paths/index.mdoc
@@ -112,8 +112,7 @@ const gridOptions = {
 };
 ```
 
-Refer to the [Aggregation](./aggregation/) page for more details, and [Editing Groups](./grouping-edit/)
-for editing aggregated values with cascading updates to children.
+Refer to the [Aggregation](./aggregation/) page for more details.
 
 ## Next Up
