+++ documentation/ag-grid-docs/src/content/docs/tree-data-row-dragging/index.mdoc
@@ -177,4 +177,3 @@ This example also demonstrates how to provide custom drop indicators using the [
 ## See Also
 
 -   [Aggregation](./aggregation/) for aggregating values in tree data
--   [Editing Groups](./grouping-edit/) for editing aggregated values with cascading updates to children
