+++ documentation/ag-grid-docs/src/content/docs/tree-data-self-referential/index.mdoc
@@ -98,8 +98,7 @@ const gridOptions = {
 };
 ```
 
-Refer to the [Aggregation](./aggregation/) page for more details, and [Editing Groups](./grouping-edit/)
-for editing aggregated values with cascading updates to children.
+Refer to the [Aggregation](./aggregation/) page for more details.
 
 ## Next Up
