+++ packages/ag-grid-community/src/edit/utils/editors.ts
@@ -5,6 +5,7 @@ import { _getCellEditorDetails } from '../../components/framework/userCompUtils'
 import type { BeanCollection } from '../../context/context';
 import type { AgColumn } from '../../entities/agColumn';
 import type { ColDef } from '../../entities/colDef';
+import type { ColDefInternal } from '../../entities/colDefInternal';
 import type { CellEditingStoppedEvent } from '../../events';
 import { _addGridCommonParams } from '../../gridOptionsUtils';
 import type {
@@ -575,7 +576,7 @@ function _columnDefsRequireValidation(columnDefs?: ColDef[]): boolean {
         return false;
     }
     for (let i = 0, len = columnDefs.length; i < len; ++i) {
-        const colDef = columnDefs[i];
+        const colDef: ColDefInternal = columnDefs[i];
         const params = colDef.cellEditorParams;
         if (!params || (!colDef.editable && !colDef.groupRowEditable)) {
             continue;
