+++ packages/ag-grid-community/src/main.ts
@@ -707,10 +707,6 @@ export {
     ValueParserParams,
     ValueSetterFunc,
     ValueSetterParams,
-    GroupRowEditableCallback,
-    GroupRowEditableCallbackParams,
-    GroupRowValueSetterParams,
-    GroupRowValueSetterFunc,
 } from './entities/colDef';
 export {
     BaseCellDataType,
