+++ packages/ag-grid-community/src/validation/rules/colDefValidations.ts
@@ -1,6 +1,7 @@
 import type { UserComponentName } from '../../context/context';
 import { _isSortDefValid, _isSortDirectionValid } from '../../entities/agColumn';
 import type { AbstractColDef, ColDef, ColGroupDef, ColumnMenuTab } from '../../entities/colDef';
+import type { ColDefInternal } from '../../entities/colDefInternal';
 import { _errMsg, toStringWithNullUndefined } from '../logging';
 import type { Deprecations, ModuleValidation, OptionsValidator, Validations } from '../validationTypes';
 import { USER_COMP_MODULES } from './userCompValidations';
@@ -39,7 +40,8 @@ export const COLUMN_DEFINITION_MOD_VALIDATIONS: ModuleValidation<ColDef | ColGro
     autoHeight: 'RowAutoHeight',
     cellClass: 'CellStyle',
     cellClassRules: 'CellStyle',
-    cellEditor: ({ cellEditor, editable, groupRowEditable }: ColDef) => {
+    cellEditor: ({ cellEditor, editable, ...rest }: ColDef) => {
+        const groupRowEditable = (rest as ColDefInternal).groupRowEditable;
         const editingEnabled = !!editable || !!groupRowEditable;
         if (!editingEnabled) {
             return null;
@@ -62,12 +64,6 @@ export const COLUMN_DEFINITION_MOD_VALIDATIONS: ModuleValidation<ColDef | ColGro
     dndSourceOnRowDrag: 'DragAndDrop',
     editable: ({ editable, cellEditor }: ColDef) => {
         if (editable && !cellEditor) {
-            return 'TextEditor';
-        }
-        return null;
-    },
-    groupRowEditable: ({ groupRowEditable, cellEditor }: ColDef) => {
-        if (groupRowEditable && !cellEditor) {
             return 'TextEditor';
         }
         return null;
@@ -260,7 +256,6 @@ const COLUMN_DEFINITION_VALIDATIONS: () => Validations<ColDef | ColGroupDef> = (
         spanRows: {
             dependencies: {
                 editable: { required: [false, undefined] },
-                groupRowEditable: { required: [false, undefined] },
                 rowDrag: { required: [false, undefined] },
                 colSpan: { required: [undefined] },
                 rowSpan: { required: [undefined] },
@@ -339,7 +334,7 @@ const COLUMN_DEFINITION_VALIDATIONS: () => Validations<ColDef | ColGroupDef> = (
     return validations;
 };
 
-type ColOrGroupKey = keyof ColDef | keyof ColGroupDef;
+type ColOrGroupKey = keyof ColDef | keyof ColGroupDef | keyof Partial<ColDefInternal>;
 const colDefPropertyMap: Record<ColOrGroupKey, undefined> = {
     headerName: undefined,
     columnGroupShow: undefined,
@@ -365,6 +360,7 @@ const colDefPropertyMap: Record<ColOrGroupKey, undefined> = {
     initialAggFunc: undefined,
     defaultAggFunc: undefined,
     aggFunc: undefined,
+    // Internal properties (not public API but still accepted)
     groupRowEditable: undefined,
     groupRowValueSetter: undefined,
     pinned: undefined,
